1.5.0-pre fold durationflipbox into flipbox
durationflipbox was almost a carbon copy of flipbox.  This adds a few bytes to flipbox
(well under 2k minimized), it was an extra 7k of code bloat in the repo.

Note that customflip will not be going away - it provides a coders reference on how these
modules are made - it's more for that use than for a true use on it's own.